% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_rows_with_pattern.R
\name{count_rows_with_pattern}
\alias{count_rows_with_pattern}
\title{Count Rows Containing a Specific Pattern in the Diamonds Dataset}
\usage{
count_rows_with_pattern(pattern)
}
\arguments{
\item{pattern}{A character string specifying the pattern to search for within the dataset rows.}
}
\value{
An integer representing the number of rows in the \code{diamonds} dataset
where the user-specified pattern is found.
}
\description{
This function writes the \code{diamonds} dataset from the \code{ggplot2} package to a temporary CSV file,
filters rows containing the user-specified pattern using a shell \code{grep} command, and counts the number
of matching rows.
}
\examples{
count_rows_with_pattern("VS")

}
