% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{get_system_info}
\alias{get_system_info}
\title{Get platform-specific system information}
\usage{
get_system_info()
}
\value{
A list with system information.
}
\description{
This function returns information about the system, including OS type.
}
\keyword{internal}
