% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{split.columns}
\alias{split.columns}
\title{Split columns based on a delimiter}
\usage{
split.columns(
  x,
  column.names = NA,
  split = ":",
  resulting.columns = 3,
  fixed = TRUE
)
}
\arguments{
\item{x}{Character vector to split}

\item{column.names}{Names for the resulting columns}

\item{split}{Delimiter to split on}

\item{resulting.columns}{Number of columns to create}

\item{fixed}{Whether to use fixed string matching}
}
\value{
data.table with split columns
}
\description{
Split columns based on a delimiter
}
